% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildup_RH_Data.R
\name{Buildup_RH_data}
\alias{Buildup_RH_data}
\title{Build up RH Data}
\usage{
Buildup_RH_data(canopy_oc_data, trait_data)
}
\arguments{
\item{canopy_oc_data}{A data frame containing canopy cover and soil organic carbon data.}

\item{trait_data}{A data frame containing plant species trait data.}
}
\value{
A standardized data frame for further analysis using Min-Max Normalization.
}
\description{
This function prepares input data for the evaluation criteria of rangeland health.
}
\details{
The function takes canopy cover, soil organic carbon (OC), and plant species trait data, and returns standardized data.

\strong{Note:} The first row of the input data matrix (canopy_oc_data) must be the reference sample, and the second column must contain the OC data, and the canopy cover must be entered as a relative value.
}
\examples{
data(canopy_oc_data)
data(trait_data)
final_data_st <- Buildup_RH_data(canopy_oc_data, trait_data)
}
