% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_RH.R
\name{Function_RH}
\alias{Function_RH}
\title{Rangeland Health}
\usage{
Function_RH(final_data_st)
}
\arguments{
\item{final_data_st}{A data frame containing standardized data from the first function.}
}
\value{
The attributes of rangeland health.
}
\description{
This function calculates attributes of rangeland health.
}
\details{
The function takes standardized data, performs predictions using pre-trained models, and returns the results.
}
\examples{
data(canopy_oc_data)
data(trait_data)
final_data_st <- Buildup_RH_data(canopy_oc_data, trait_data)
evaluation.criteria <- Function_RH(final_data_st)
}
