% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_input_model.R
\name{plot_input_model}
\alias{plot_input_model}
\title{plot input model}
\arguments{
\item{final_data_st}{A data frame containing standardized data from the first function.}

\item{row_indices}{A vector of row indices specifying which rows to use for creating sample plots.}

\item{plot_title_prefix}{A prefix for the plot titles (default is "Site").}

\item{ncol}{Number of columns for arranging the plots. Default is 3.}
}
\value{
A list of sample charts.
}
\description{
This function creates graphs for the inputs of the Function_RH model.
}
\details{
The function takes input data and specific row indices, creates sample charts, and returns the plots.
}
\examples{
data(canopy_oc_data)
data(trait_data)
final_data_st <- Buildup_RH_data(canopy_oc_data, trait_data)
row_indices <- 1:17
plot_input_model <- plot_input_model(
  final_data_st,
  row_indices,
  plot_title_prefix = "Site",
  ncol = 7,
  font_family = "Arial"
)
}
