% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RH.R
\name{plot_RH}
\alias{plot_RH}
\title{Create Evaluation Criteria Plots}
\arguments{
\item{evaluation.criteria}{A data frame containing standardized data from the first function.}

\item{selected_columns}{A vector of column indices specifying which criteria to plot. If NULL, all columns will be plotted.}

\item{ncol}{Number of columns for arranging the plots. Default is 4.}

\item{plot_RHC}{A function that takes standardized data and generates evaluation plots for different criteria of rangeland health.}
}
\value{
A list of evaluation criteria(attributes) plots.
}
\description{
This function creates plots to compare evaluation criteria of rangeland health.
}
\details{
The function takes the output from Function_RH and creates evaluation plots for different criteria.
}
\examples{
data(canopy_oc_data)
data(trait_data)
final_data_st <- Buildup_RH_data(canopy_oc_data, trait_data)
evaluation.criteria <- Function_RH(final_data_st)
# Plot all columns
plots_all <- plot_RH(evaluation.criteria, ncol = 4)
# Plot specific columns
selected_columns <- c(1, 3)
plots_selected <- plot_RH(evaluation.criteria, selected_columns, ncol = 2)
}
